+++ src/cmakeProject.ts
@@ -926,7 +926,8 @@ export class CMakeProject {
                     }
                     if (selectedFile) {
                         const newSourceDirectory = path.dirname(selectedFile);
-                        await this.setSourceDir(await util.normalizeAndVerifySourceDir(newSourceDirectory, CMakeDriver.sourceDirExpansionOptions(this.workspaceContext.folder.uri.fsPath)));
+                        await this.setSourceDir(await util.normalizeAndVerifySourceDir(newSourceDirectory,
+                            await CMakeDriver.sourceDirExpansionOptions(this.workspaceContext.folder.uri.fsPath, this.useCMakePresets, this.getActiveKit())));
                         void vscode.workspace.getConfiguration('cmake', this.workspaceFolder.uri).update("sourceDirectory", this._sourceDir);
                         if (config) {
                             // Updating sourceDirectory here, at the beginning of the configure process,
@@ -1142,7 +1143,11 @@ export class CMakeProject {
      */
     private async init(sourceDirectory: string) {
         log.debug(localize('second.phase.init', 'Starting CMake Tools second-phase init'));
-        await this.setSourceDir(await util.normalizeAndVerifySourceDir(sourceDirectory, CMakeDriver.sourceDirExpansionOptions(this.workspaceContext.folder.uri.fsPath)));
+        this.kitsController = await KitsController.init(this);
+        this.presetsController = await PresetsController.init(this, this.kitsController, this.isMultiProjectFolder);
+        await this.doUseCMakePresetsChange();
+        await this.setSourceDir(await util.normalizeAndVerifySourceDir(sourceDirectory,
+            await CMakeDriver.sourceDirExpansionOptions(this.workspaceContext.folder.uri.fsPath, this.useCMakePresets, this.getActiveKit())));
         this.doStatusChange(this.workspaceContext.config.options);
         // Start up the variant manager
         await this.variantManager.initialize(this.folderName);
@@ -1168,11 +1173,6 @@ export class CMakeProject {
         this.cPackageController.onPackagingEnabledChanged(enabled => this._cpackEnabled.set(enabled));
 
         this.statusMessage.set(localize('ready.status', 'Ready'));
-
-        this.kitsController = await KitsController.init(this);
-        this.presetsController = await PresetsController.init(this, this.kitsController, this.isMultiProjectFolder);
-
-        await this.doUseCMakePresetsChange();
 
         this.disposables.push(this.onPresetsChanged(() => this.doUseCMakePresetsChange()));
         this.disposables.push(this.onUserPresetsChanged(() => this.doUseCMakePresetsChange()));
@@ -1239,6 +1239,27 @@ export class CMakeProject {
         return this.presetsController.onUserPresetsChanged(listener);
     }
 
+    /**
+     * Calculate which kit object should be used for variable exapansion.
+     * Obviously, the active kit if set but otherwise, since variable expansion may need kit information before the active kit is determined
+     * (like very early during project load) then deduce from previous user selection which was saved in the workspace state.
+     * Do not change any flow related to how and when this.activeKit is set. When it is null, this method just returns a kit (if we find one)
+     * but without setting it as active here.
+     */
+    public getActiveKit(): Kit | null {
+        if (this.activeKit) {
+            return this.activeKit;
+        }
+
+        const kitName: string | null = this.workspaceContext.state.getActiveKitName(this.folderName, this.isMultiProjectFolder);
+        if (kitName) {
+            // It remembers a kit. Find it in the kits avail in this dir:
+            return this.kitsController.availableKits.find(k => k.name === kitName) || null;
+        }
+
+        return null;
+    }
+
     async initializeKitOrPresets() {
         if (this.useCMakePresets) {
             const latestConfigPresetName = this.workspaceContext.state.getConfigurePresetName(this.folderName, this.isMultiProjectFolder);
@@ -1252,10 +1273,8 @@ export class CMakeProject {
             }
         } else {
             // Check if the CMakeProject remembers what kit it was last using in this dir:
-            const kitName = this.workspaceContext.state.getActiveKitName(this.folderName, this.isMultiProjectFolder);
-            if (kitName) {
-                // It remembers a kit. Find it in the kits avail in this dir:
-                const kit = this.kitsController.availableKits.find(k => k.name === kitName) || null;
+            const kit = this.getActiveKit();
+            if (kit) {
                 // Set the kit: (May do nothing if no kit was found)
                 await this.setKit(kit);
             }